+++ server/src/main/java/server/team33/item/entity/Brand.java
@@ -1,7 +1,14 @@
 package server.team33.item.entity;
 
 public enum Brand {
-    BRAND1,
-    BRAND2,
-    BRAND3
+    MYNI,
+    Centrum,
+    Habitual,
+    NaturalLife,
+    Roniewell,
+    GNM,
+    nzOrigin,
+    FarmAndTop,
+    NutrioneLife,
+    SOLGAR
 }
