+++ server/src/main/java/server/team33/user/mapper/UserMapper.java
@@ -41,7 +41,7 @@ public interface UserMapper {
                 .email(user.getEmail())
                 .realName(user.getRealName())
                 .updatedAt(method == HttpMethod.GET ? user.getCreatedAt() : ZonedDateTime.now(ZoneId.of("Asia/Seoul")))
-                .password(method == HttpMethod.GET ? null : "비밀번호가 변경되었습니다!")
+                .password("정보가 변경되었습니다.")
                 .social(!user.getProviderId().equals("default"))
                 .build();
     }
